# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VisiPoint V5 API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/Users/{entity_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Users",
                "description": "Users",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Pass User Details",
                    "required": true
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Users listed Successfully"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "xxxxxx"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/Users/{User}": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Edit User",
                "description": "Edit User",
                "operationId": "User",
                "requestBody": {
                    "description": "Pass User Details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "User": {
                                        "properties": {
                                            "first_name": {
                                                "type": "string",
                                                "example": "Foo"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Bar"
                                            },
                                            "User_type_id": {
                                                "type": "string",
                                                "example": "User1@mail.com"
                                            },
                                            "system_User": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "rfid_number": {
                                                "type": "string",
                                                "example": "123456"
                                            },
                                            "User_id": {
                                                "type": "string",
                                                "example": "123456"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "iamgeUrl"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "test@gmail.com"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "0127931"
                                            },
                                            "role_id": {
                                                "type": "string",
                                                "example": "1111"
                                            },
                                            "department_id": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "PassWord12345"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "entity": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "1234"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Users Updated Successfully"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "xxxxxx"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}